+++ quackosm/__init__.py
@@ -9,7 +9,7 @@ from quackosm.functions import convert_pbf_to_gpq, get_features_gdf
 from quackosm.pbf_file_reader import PbfFileReader
 
 __app_name__ = "QuackOSM"
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 __all__ = [
     "PbfFileReader",