+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2024-01-06
+
 ### Added
 
 - Speed column for Rich progress bar
@@ -45,7 +47,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created QuackOSM repository
 - Implemented PbfFileReader
 
-[Unreleased]: https://github.com/kraina-ai/quackosm/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/kraina-ai/quackosm/compare/0.3.1...HEAD
+
+[0.3.1]: https://github.com/kraina-ai/quackosm/compare/0.3.0...0.3.1
 
 [0.3.0]: https://github.com/kraina-ai/quackosm/compare/0.2.0...0.3.0